refactor(origins): refactor aspects to array
