+++ extensions/origins/source/aspects/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const http = require('./http')
+
+module.exports = [http.Aspect]
