+++ features/steps/.workspace/components/collection/origins.amqp/operations/test.js
@@ -3,6 +3,7 @@
 async function observation (input, none, context) {
   const event = { happened: true }
 
+  await context.origins.amqp.invoke('emit', 'test_event', event)
   await context.amqp.queue.emit('test_event', event)
 }
